Add a bunch of repo tests
