+++ GlobalUsings.cs
@@ -1,0 +1,8 @@
+// Adding this file to .csproj from:
+//https://ingebrigtsen.blog/2021/08/13/c-10-reuse-global-usings-in-multiple-projects/
+
+// Namespaces
+// global using global::SimpleDB;
+// global using global::Chirp.Infrastructure;
+
+// 3rd-party class libraries
