+++ test/Chirp.Razor.Tests/RazorAPITests.cs
@@ -1,0 +1,18 @@
+namespace Chirp.Razor.Tests;
+using Xunit;
+
+public class RazorAPITests
+{
+    /*
+    [Fact]
+    public void CheckIfBothStartingCheepsAreDisplayed() 
+    {
+        //Arrange
+
+        //Act
+        
+
+        //Assert
+
+    }*/
+}
